MVC
In this demo project, there is a list of products with their titles, ratings, and images. Prices are also included, and the detail page displays all the relevant information. The design for adding products to the cart is implemented, and product descriptions are available.

Firebase is integrated for login and sign-up functionality. Users can sign up with Firebase, and there is an option to log in. Additionally, a feature for marking products as favorites has been added. Clicking on the favorite option adds the product to the cart. For demonstration purposes, new users can also sign up, and Firebase credentials for email authentication have been provided.

Please review and provide feedback on the project.